refactor(uds): rehome the run-boundary record types in the engine
Phase 3a of the format-blind extraction: the data the session holds —
the interface-file records (routing, RDII, runoff replay, rain gage
and station readings) and the §15.8 overland ledger row — moves into
the engine proper (simulation::records and the overland module), while
the §14.8 grammars that read and write the predecessor's files stay in
io and import the types. Historical io:: paths remain as re-export
shims until the lift re-points external consumers; every in-engine
consumer is re-pointed now. What remains of the engine's io references
is exactly the phase-3 break set: the text-opening family and the
results, report and sidecar attach points.
