+++ crates/engine-uds/src/simulation/mod.rs
@@ -14,6 +14,7 @@ pub mod checkpoint;
 pub mod controls;
 pub mod coupled;
 pub mod expression;
+pub mod records;
 pub mod time;
 
 pub mod engine;
